Ajout de la sécurité

+++ backend/controllers/comment.js
@@ -38,9 +38,11 @@ exports.modifyComment = (req, res, next) => {
         else if (!req.body.content) {
             return res.status(400).json({ message: "Veuillez ne pas laisser les champs vides !"})
         }
+        else if (comment.userId === req.auth.userId || req.auth.isAdmin) {
         comment.update({ content: req.body.content })
         .then(() => res.status(200).json({ message: "Commentaire modifié !" }))
         .catch((error) => res.status(400).json({error}))
+        }
        /* Comment.update({ where: { id: req.params.id} }), ( {content: req.body.content}, { where: { id: req.params.id} })
         .then(() => res.status(200).json({ message: 'Commentaire modifié !'}))
         .catch(error => res.status(400).json({error}))*/
@@ -64,9 +66,11 @@ exports.deleteComment = (req, res, next) => {
                 error: new Error('Requête non autorisée !')
             }); 
         }*/
+        if (comment.userId === req.auth.userId || req.auth.isAdmin) {
         comment.destroy({ where: { id: req.params.id} })
         .then(() => res.status(200).json({ message: 'Commentaire supprimé !'}))
         .catch(error => res.status(400).json({error}))
+        }
     })
     .catch(error => res.status(500).json({ error}));
 };
